requirements.txt
customtkinter>=5.2
Pillow>=10.3
